% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestation_data.R
\docType{data}
\name{firestation}
\alias{firestation}
\title{Korea firestation data}
\format{An object of class \code{data.frame} with 215 rows and 6 columns.}
\source{
\href{https://www.data.go.kr/}{Korea Public Data Portals}
}
\usage{
data(firestation)
}
\description{
This data include national firestation address, telephone number, longitude
and latitude
}
\examples{
data(firestation)
}
\keyword{datasets}
