% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_kakao.R
\name{geocode_kakao}
\alias{geocode_kakao}
\title{Geocode from KAKAO}
\usage{
geocode_kakao(address, kakao_key)
}
\arguments{
\item{address}{address that you want to know it`s longitude and latitude}

\item{kakao_key}{you must input your KAKAO REST API key}
}
\value{
longitude and latitude
}
\description{
geocode_kakao function will get longitude and latitude from KAKAO api.
}
\examples{
  geocode_kakao('address', kakao_key)
}
