% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_keyword_kakao.R
\name{geocode_keyword_kakao}
\alias{geocode_keyword_kakao}
\title{Geocode from KAKAO using keyword}
\usage{
geocode_keyword_kakao(keyword, kakao_key, category_group_code, lon, lat,
  radius, rect)
}
\arguments{
\item{keyword}{it use nomination that you want to know it`s latitude and longitude}

\item{kakao_key}{you must input your KAKAO REST API key}

\item{category_group_code}{tou can input this to narrow range}

\item{radius}{you can input this to narrow range and the center of range is lon and lat}
}
\value{
latitude and longitude
}
\description{
geocode_keyword_kakao will get longitude and latitude from KAKAO api.
it will use nomination or building name
it will narrow range by category_group_code or radius.
}
