% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_naver.R
\name{geocode_naver}
\alias{geocode_naver}
\title{Geocode from NAVER}
\usage{
geocode_naver(address, naver_key, naver_secret)
}
\arguments{
\item{address}{address that you want to know it`s longitude and latitude}

\item{naver_key}{you must input your NAVER client id}

\item{naver_secret}{you must input your NAVER client secret}
}
\value{
longitude and latitude
}
\description{
geocode_naver function will get longitude and latitude from NAVER api.
}
