% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kakao_navermap.R
\name{get_kakao_navermap}
\alias{get_kakao_navermap}
\title{Get NAVER map and use KAKAO api}
\usage{
get_kakao_navermap(center = c(lon = 126.9849208, lat = 37.5664519),
  zoom = 4, size = c(640, 640), format = c("png", "jpeg", "jpg"),
  crs = c("EPSG:4326", "NHN:2048", "NHN:128", "EPSG:4258", "EPSG:4162",
  "EPSG:2096", "EPSG:2097", "EPSG:2098", "EPSG:900913"),
  baselayer = c("default", "satellite"), color = c("color", "bw"),
  overlayers = c("anno_satellite", "bicycle", "roadview", "traffic"),
  markers, markers_ab, naver_key, url_2 = "http://naver.com",
  keyword = FALSE, address = FALSE, kakao_key, naver_secret,
  category_group_code, filename = "ggmapTemp", messaging = FALSE,
  urlonly = FALSE, force = FALSE, where = tempdir(),
  archiving = TRUE, ...)
}
\arguments{
\item{center}{the center of map, you can use diameter, nomination or address}

\item{zoom}{map zoom, it is 1~14 integer}

\item{size}{rectangular dimension in map}

\item{format}{image format}

\item{crs}{"EPSG:4326" is default}

\item{baselayer}{you can select "satelite" or "default"}

\item{markers}{data.frame with first column longitude, second column latitude, markers displayed in map}

\item{markers_ab}{same as markers but it will show alphabets}

\item{naver_key}{NAVER api client id}

\item{url_2}{NAVER register url, default is http://naver.com. you should input your url if you have different url}

\item{keyword}{if you use nomination then you input TRUE}

\item{address}{if you use address then you input TRUE, then you should input keyword FALSE, when you use diameter you should input FALSE for keyword and address}

\item{kakao_key}{KAKAO REST api key}

\item{naver_secret}{NAVER api client secret}

\item{category_group_code}{you can narrow range you can reference code at https://github.com/subinjo92/nkmap}

\item{messaging}{turn messaging on.off}

\item{urlonly}{return url only}

\item{force}{if the map is on local file, should a new map be looked up or not}

\item{where}{where shoul the tile drawer be located}

\item{archiving}{use archived maps by changing to TRUE you agree to abide by any of the rules governing caching naver maps}

\item{colors}{color or black-white}

\item{overlayer}{you can use "anno_satellite","bicycle", "roadview", "traffic".}
}
\value{
NAVER static map
}
\description{
get_kakao_navermap will get NAVER Static map and it use NAVER Static map api.
you can use diameter, nomination or address.
you can use this function when you have id and secret
https://developers.naver.com/main/
}
\examples{
  ggmap(get_kakao_navermap(center = c(lon = 126, lat = 35), zoom = 3, naver_key = naver_key, url_2 = url_2, keyword = F, address = F, kakao_key = kakao_key, naver_secret = naver_secret, baselayer ='satellite', overlayer = c("anno_satellite", 'traffic')))
  ggmap(get_kakao_navermap(center = 'address', zoom = 10, naver_key = naver_key, keyword = F, address = T, kakao_key = kakao_key, naver_secret = naver_secret, baselayer ='satellite', overlayer = c("anno_satellite", 'traffic')))
  ggmap(get_kakao_navermap(center = 'nomination', zoom = 10, naver_key = naver_key, url_2 = url_2, keyword = T, address = F, kakao_key = kakao_key, naver_secret = naver_secret, baselayer ='satellite', overlayer = c("anno_satellite", 'traffic')))
}
