% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_package.R
\name{use_package}
\alias{use_package}
\title{Use package more easily}
\usage{
use_package(package, use = TRUE)
}
\arguments{
\item{package}{package name that you want to install and require}

\item{use}{you want to require or not, if TRUE it will require the package}
}
\description{
when you use this function you can do install and require at once
for installed packages, it only do require
}
