% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestation_data.R
\docType{data}
\name{firestation}
\alias{firestation}
\title{Korea firestation data}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://www.data.go.kr/}{Korea Public Data Portals}
}
\usage{
data(firestation)
}
\description{
This data include national firestation address, telephone number, longitude
and latitude
}
\examples{
data(firestation)
\donttest{
iplotCurves(phe, times, phe[,c(61,121)], phe[,c(121,181)],
            chartOpts=list(curves_xlab="Time (hours)", curves_ylab="Root tip angle (degrees)",
                           scat1_xlab="Angle at 2 hrs", scat1_ylab="Angle at 4 hrs",
                           scat2_xlab="Angle at 4 hrs", scat2_ylab="Angle at 6 hrs"))}
}
\keyword{datasets}
